requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
transformers>=4.39.3
torch==2.1.1
torchvision==0.16.1
accelerate==0.25.0
mlx>=0.29.1
mlx-lm>=0.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.7
langgraph>=0.0.20
pyttsx3==2.90
speechrecognition==3.10.0
pyaudio==0.2.11
websockets==12.0
python-socketio>=5.8.0
aiofiles==23.2.1
httpx==0.25.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
psutil==5.9.6
loguru==0.7.2
prometheus-client==0.19.0
opentelemetry-api>=1.22.0
redis>=4.5.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
# Security Enhancements - Phase 1
slowapi>=0.1.9
email-validator>=2.1.0
# Performance Enhancements - Phase 2
aiofiles>=23.2.1
aiomultiprocess>=0.9.0
psutil>=5.9.6
memory-profiler>=0.61.0
aiocache>=0.12.0
# PAI Implementation - Complete PAI System
anthropic>=0.7.0         # Claude integration
openai>=1.0.0            # GPT integration
google-generativeai>=0.4.0 # Gemini
requests>=2.31.0         # HTTP client for MCP servers
playwright>=1.40.0       # Browser automation MCP
selenium>=4.15.0         # Web testing MCP
stripe>=7.0.0            # Financial MCP
squareup>=14.0.0         # Payment MCP
matplotlib>=3.8.2        # Data visualization
plotly>=5.17.0           # Interactive charts
beautifulsoup4>=4.12.0   # Web scraping
yfinance>=0.2.0          # Financial data
tweepy>=4.14.0           # Twitter/social integration
discord.py>=2.3.0        # Discord integration
twilio>=8.2.0            # SMS integration
sendgrid>=6.10.0         # Email integration
jupyter>=1.0.0           # Notebook support
nbclient>=0.8.0          # Execute notebooks
schedule>=1.2.1          # Task scheduling
apscheduler>=3.10.0      # Advanced scheduling
regex>=2023.10.0         # Advanced regex patterns
filelock>=3.13.0         # File locking for UFC context

# Code Quality Enhancements - Phase 3
mypy>=1.7.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
pylint>=3.0.0
pre-commit>=3.5.0
pytest-cov>=4.1.0
hypothesis>=6.87.0
tox>=4.11.0
python-multipart==0.0.6
cryptography==41.0.7
pytest>=7.0.0
pytest-asyncio>=0.21.0
black==23.11.0
ipython>=8.0.0
click==8.1.7
rich==13.7.0
python-dotenv==1.0.0
jinja2==3.1.2
schedule==1.2.1
python-dateutil==2.8.2
distro==1.8.0

# Latest Advanced Agent Orchestration Frameworks (Phase 1 & 2)
langgraph>=0.26.0      # Latest LangGraph for state management & orchestration
langchain-core>=0.3.0  # Latest LangChain core for agent interactions
langchain>=0.3.0       # Latest LangChain for tool integrations
autogen>=0.4.0         # Latest Microsoft AutoGen for multi-agent conversations
crewai>=0.70.0         # Latest CrewAI for agent team management (formerly crewai)
smolagents>=1.0.0      # Latest SmolAgents for lightweight agent creation

# Advanced Multi-Agent Frameworks (Phase 3)
superagi>=0.1.0        # Self-evolving agent platform
autogpt>=0.5.0         # Latest AutoGPT for autonomous task execution
babyagi>=1.0.0         # Latest BabyAGI for task decomposition

# Modern Agent Communication & State Management
astrapy>=1.5.0         # Astra DB for vector storage (if needed)
faiss-cpu>=1.8.0       # Latest FAISS for vector similarity
weaviate-client>=4.0.0 # Weaviate for advanced vector search
qdrant-client>=1.9.0   # Qdrant for distributed vector storage

# Advanced Orchestration & Planning
plan-and-solve>=0.1.0  # Task planning and solving framework
reasoning-engine>=0.1.0 # Advanced reasoning capabilities

# Agent Communication & Coordination
websockets>=12.0       # Real-time agent communication
aio-pika>=9.0.0        # Message queue for agent coordination
redis>=4.5.0           # Shared state for agent coordination
python-multipart       # Multipart data handling for agents
